Move functions and variaables for playback into a new script. Made some stuff public to facilitate.
